I defined the spinor function that we will need for for the psi
In the script I added the spinor function of the electron before
it goes through the magnetic field. The spinor function in is the
inverse fourier transform of the gaussian defined in the text file.
Also I added a pdf of the plot of the spinor function where the function
is centered around zero.